new session file created, globalvars made as dictionary
