+++ python/grass/grassdb/globals.py
@@ -1,0 +1,17 @@
+"""
+Set global variables for objects in a GRASS GIS Spatial Database
+
+(C) 2020 by the GRASS Development Team
+This program is free software under the GNU General Public
+License (>=v2). Read the file COPYING that comes with GRASS
+for details.
+
+.. sectionauthor:: Linda Kladivova <linda.kladivova seznam cz>
+"""
+
+# global variables
+globalvars = {}
+globalvars["DEFAULT_LOCATION"] = "world_latlong_wgs84"
+globalvars["TEMPORARY_LOCATION"] = "tmploc"
+globalvars["UNKNOWN_LOCATION"] = "<UNKNOWN>"
+globalvars["UNKNOWN_MAPSET"] = "<UNKNOWN>"
