--- python/grass/grassdb/strings.py
@@ -1,16 +1,0 @@
-"""
-Set strings for objects in a GRASS GIS Spatial Database
-
-(C) 2020 by the GRASS Development Team
-This program is free software under the GNU General Public
-License (>=v2). Read the file COPYING that comes with GRASS
-for details.
-
-.. sectionauthor:: Linda Kaldivova <linda.kladivova seznam cz>
-"""
-
-# constant strings
-default_location = "world_latlong_wgs84"
-temporary_location = "tmploc"
-unknown_location = "<UNKNOWN>"
-unknown_mapset = "<UNKNOWN>"
